refactor(file-tree): kebab-case file and directory names
Renames the FileTree folder and its sources to kebab-case
(file-tree/, directory-row.tsx, file-row.tsx, file-tree.test.tsx)
and updates the two import sites.
