+++ frontend/src/components/file-tree/directory-row.tsx
