+++ frontend/src/components/file-tree/file-row.tsx
